Fix up some typings
Pandas doesn't have a stub file, and I was getting tired of warnings from mypy about it.

Mypy also seems to complain about the zip() within the sensor filter list comprehension. Seems to be something to do with the usage of "in", since the rest of them are fine. Error was that there was no matching of any overload variant of zip() that would give List[bool]. But the any() part of the comprehension takes care of that.