Reduce complexity as scored by flake8
The function parse_csv() had a complexity score of 13, largely due to the filtering portion of it. Moving that code into its own function, _filter_csv(), brought it back under 10.

But now, _filter_csv() has a complexity score of 11. Since it's so close to 10, it's just silenced for now. But a future todo is to actually reduce it to under 10.